Committing rest of tbc_parser, but...
Since it relies on actual nodes for looking up scope gates, we're unable to do that when a cursor is on some whitespace for which there is no node. For example, in `class Foo; end`, when the cursor is between the `;` and `end`, since there's no AST node there, the only scope that's registered is Foo's parent, which happens to be the root. I think I could make this work, but when coupled with the growing notion that I probably don't actually want to use lib-ruby-parser (since it's geared towards interpreting Ruby, not editing Ruby code in an IDE), I don't think I want to spend more time on this approach.